chore: Update Supabase dependencies and improve login/signup form styling
This commit updates the Supabase dependencies in the frontend/package.json file to include the latest version of "@supabase/supabase-js" (v2.44.4). This update ensures that the project uses the most up-to-date features and bug fixes provided by Supabase.

Additionally, the commit improves the styling and layout of the login and signup forms. In Login.css, the padding for the login form button is commented out, removing the padding. In Signup.css, the padding for the signup form is increased to 30px, and a fixed width of 360px is set for the form. The text-align property is also set to center for the form and the h2 heading. These changes enhance the visual presentation and alignment of the signup form, improving the user experience and visual consistency of the login and signup pages.