chore: Update .gitignore to include backend/imgs/*
This commit updates the .gitignore file to include the pattern "backend/imgs/*". This ensures that all files and directories within the "backend/imgs" directory are ignored by Git and not tracked in the repository.

These changes improve the repository's cleanliness and ensure that only necessary files are included while ignoring irrelevant files.